+++ Drivers/Common/Src/drivers_config.cpp
@@ -5,7 +5,6 @@
 */
 SBUSReceiver sbus_instance(sbus_uart);
 SBUSReceiver* sbus_pointer = &sbus_instance;
-    creating RFD900 instance
 
 UARTDevice rfd900Instance(RFD900_UART);
 UARTDevice* pRFD900 = &rfd900Instance;
